requirements.txt
Django==4.2.16
djangorestframework==3.15
psycopg2>=2.0.6,>2.9
drf-spectacular==0.27.1
django-cors-headers==4.0.0
